requirements.txt
absl-py==2.1.0
ale-py==0.8.1
arrow==1.3.0
attrs==23.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
boto3==1.34.78
botocore==1.34.78
box2d-py==2.3.5
bravado==11.0.3
bravado-core==6.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
decorator==4.4.2
docopt==0.6.2
etils==1.8.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.13.3
flatten-dict==0.4.2
fonttools==4.50.0
fqdn==1.5.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
gymnasium==0.29.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_resources==6.4.0
isoduration==20.11.0
jax==0.4.26
jaxlib==0.4.26
Jinja2==3.1.3
jmespath==1.0.1
jsonpickle==3.0.3
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.8.4
ml-dtypes==0.4.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mujoco==3.1.3
mujoco-py==2.1.2.14
munch==4.0.0
neptune==1.10.1
neptune-sacred==1.0.1
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
proglog==0.1.10
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.22
pygame==2.5.2
PyJWT==2.8.0
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
s3transfer==0.10.1
sacred==0.8.5
scipy==1.13.0
Shimmy==0.2.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
swagger-spec-validator==3.0.3
swig==4.2.1
sympy==1.12
torch==2.2.2
tqdm==4.66.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
webcolors==1.13
websocket-client==1.7.0
wrapt==1.16.0
zipp==3.18.1
